+++ ui/src/components/date-time-input/QasDateTimeInput.vue
@@ -1,19 +1,21 @@
 <template>
-  <qas-input ref="input" v-bind="attributes" v-model="currentValue" :unmasked-value="false" @update:model-value="updateModelValue">
-    <template #append>
-      <q-icon v-if="!timeOnly" class="cursor-pointer" name="o_event">
-        <q-popup-proxy ref="dateProxy" transition-hide="scale" transition-show="scale">
-          <q-date v-model="currentValue" v-bind="dateProps" :mask="maskDate" @update:model-value="updateModelValue" />
-        </q-popup-proxy>
-      </q-icon>
+  <div>
+    <qas-input ref="input" v-bind="attributes" v-model="currentValue" :unmasked-value="false" @update:model-value="updateModelValue">
+      <template #append>
+        <q-icon v-if="!timeOnly" class="cursor-pointer" name="o_event">
+          <q-popup-proxy ref="dateProxy" transition-hide="scale" transition-show="scale">
+            <q-date v-model="currentValue" v-bind="dateProps" :mask="maskDate" @update:model-value="updateModelValue" />
+          </q-popup-proxy>
+        </q-icon>
 
-      <q-icon v-if="!dateOnly" class="cursor-pointer q-ml-md" name="o_access_time">
-        <q-popup-proxy ref="timeProxy" transition-hide="scale" transition-show="scale">
-          <q-time v-model="currentValue" v-bind="timeProps" format24h :mask="maskDate" @update:model-value="updateModelValue" />
-        </q-popup-proxy>
-      </q-icon>
-    </template>
-  </qas-input>
+        <q-icon v-if="!dateOnly" class="cursor-pointer q-ml-md" name="o_access_time">
+          <q-popup-proxy ref="timeProxy" transition-hide="scale" transition-show="scale">
+            <q-time v-model="currentValue" v-bind="timeProps" format24h :mask="maskDate" @update:model-value="updateModelValue" />
+          </q-popup-proxy>
+        </q-icon>
+      </template>
+    </qas-input>
+  </div>
 </template>
 
 <script>
@@ -130,6 +132,7 @@ export default {
       const valueLength = value.replace(/_/g, '').length
 
       if (value === '' || valueLength === this.mask.length) {
+        // console.log(value, '>>>>> value')
         this.lastValue = this.timeOnly ? value : this.toISOString(value)
         this.$emit('update:modelValue', this.lastValue)
       }