+++ ui/src/components/field/QasField.vue
@@ -49,7 +49,7 @@ export default {
     },
 
     modelValue: {
-      default: () => ({}),
+      default: undefined,
       type: [Object, Array, String, Number, Boolean]
     }
   },
@@ -189,6 +189,8 @@ export default {
       if (!this.isBoolean) {
         return false
       }
+
+      console.log(this.modelValue, '>>>> this.modelValue')
 
       return !this.modelValue && (this.modelValue === undefined || typeof this.modelValue === 'string')
     },