+++ ui/src/vue-plugin.js
@@ -14,12 +14,16 @@ import QasDateTimeInput from './components/date-time-input/QasDateTimeInput.vue'
 import QasDebugger from './components/debugger/QasDebugger.vue'
 import QasDialog from './components/dialog/QasDialog.vue'
 import QasDialogRouter from './components/dialog-router/QasDialogRouter.vue'
+import QasField from './components/field/QasField.vue'
+import QasFormGenerator from './components/form-generator/QasFormGenerator.vue'
+import QasFormView from './components/form-view/QasFormView.vue'
 import QasGallery from './components/gallery/QasGallery.vue'
 import QasGridGenerator from './components/grid-generator/QasGridGenerator.vue'
 import QasInput from './components/input/QasInput.vue'
 import QasLabel from './components/label/QasLabel.vue'
 import QasLayout from './components/layout/QasLayout.vue'
 import QasListItems from './components/list-items/QasListItems.vue'
+import QasNestedFields from './components/nested-fields/QasNestedFields.vue'
 import QasNumericInput from './components/numeric-input/QasNumericInput.vue'
 import QasPasswordInput from './components/password-input/QasPasswordInput.vue'
 import QasPasswordStrengthChecker from './components/password-strength-checker/QasPasswordStrengthChecker.vue'
@@ -27,16 +31,13 @@ import QasSearchBox from './components/search-box/QasSearchBox.vue'
 import QasSelect from './components/select/QasSelect.vue'
 import QasSelectList from './components/select-list/QasSelectList.vue'
 import QasSignaturePad from './components/signature-pad/QasSignaturePad.vue'
+import QasSignatureUploader from './components/signature-uploader/QasSignatureUploader.vue'
 import QasSortable from './components/sortable/QasSortable.vue'
 import QasTableGenerator from './components/table-generator/QasTableGenerator.vue'
 import QasTabsGenerator from './components/tabs-generator/QasTabsGenerator.vue'
 import QasTextTruncate from './components/text-truncate/QasTextTruncate.vue'
 import QasTransfer from './components/transfer/QasTransfer.vue'
 import QasUploader from './components/uploader/QasUploader.vue'
-import QasSignatureUploader from './components/signature-uploader/QasSignatureUploader.vue'
-import QasField from './components/field/QasField.vue'
-import QasFormGenerator from './components/form-generator/QasFormGenerator.vue'
-import QasNestedFields from './components/nested-fields/QasNestedFields.vue'
 
 import {
   Dialog,
@@ -63,12 +64,16 @@ function install (app) {
   app.component('QasDebugger', QasDebugger)
   app.component('QasDialog', QasDialog)
   app.component('QasDialogRouter', QasDialogRouter)
+  app.component('QasField', QasField)
+  app.component('QasFormGenerator', QasFormGenerator)
+  app.component('QasFormView', QasFormView)
   app.component('QasGallery', QasGallery)
   app.component('QasGridGenerator', QasGridGenerator)
   app.component('QasInput', QasInput)
   app.component('QasLabel', QasLabel)
   app.component('QasLayout', QasLayout)
   app.component('QasListItems', QasListItems)
+  app.component('QasNestedFields', QasNestedFields)
   app.component('QasNumericInput', QasNumericInput)
   app.component('QasPasswordInput', QasPasswordInput)
   app.component('QasPasswordStrengthChecker', QasPasswordStrengthChecker)
@@ -76,16 +81,13 @@ function install (app) {
   app.component('QasSelect', QasSelect)
   app.component('QasSelectList', QasSelectList)
   app.component('QasSignaturePad', QasSignaturePad)
+  app.component('QasSignatureUploader', QasSignatureUploader)
   app.component('QasSortable', QasSortable)
   app.component('QasTableGenerator', QasTableGenerator)
   app.component('QasTabsGenerator', QasTabsGenerator)
   app.component('QasTextTruncate', QasTextTruncate)
   app.component('QasTransfer', QasTransfer)
   app.component('QasUploader', QasUploader)
-  app.component('QasSignatureUploader', QasSignatureUploader)
-  app.component('QasField', QasField)
-  app.component('QasFormGenerator', QasFormGenerator)
-  app.component('QasNestedFields', QasNestedFields)
 
   app.config.globalProperties.$qas = {
     error: NotifyError,
@@ -114,12 +116,16 @@ export {
   QasDebugger,
   QasDialog,
   QasDialogRouter,
+  QasField,
+  QasFormGenerator,
+  QasFormView,
   QasGallery,
   QasGridGenerator,
   QasInput,
   QasLabel,
   QasLayout,
   QasListItems,
+  QasNestedFields,
   QasNumericInput,
   QasPasswordInput,
   QasPasswordStrengthChecker,
@@ -127,16 +133,13 @@ export {
   QasSelect,
   QasSelectList,
   QasSignaturePad,
+  QasSignatureUploader,
   QasSortable,
   QasTableGenerator,
   QasTabsGenerator,
   QasTextTruncate,
   QasTransfer,
   QasUploader,
-  QasSignatureUploader,
-  QasField,
-  QasFormGenerator,
-  QasNestedFields,
 
   // plugin
   Dialog,